Use internal per-socket epoll instance to hide transport fds
Instead of exposing all transport fds directly to the application, put
them in a per-socket epoll instance. This change allows the
introduction of two new functions, xcm_await() and xcm_fd(), which
render xcm_want() obsolete and greatly simplifies using XCM sockets in
a single-threaded, event-driven program.

The XCM API/ABI is now version 13, and remains backward
compatible. The xcm_want() function and some related defines are moved
to <xcm_compat.h>. The documentation is updated to focus on the two
new functions.

Legacy applications, still using xcm_want(), will be given the same fd
- that of the epoll instance - with the same event (XCM_FD_READABLE)
in every xcm_want() call.

XCM now promises that fd numbers will not be reused (for new file
objects) between xcm_want() invocations. This is relevant for
applications using an epoll-based event loop.

Performance stays about the same after this patch, except for TLS
connection sockets, which suffers a slight degradation. However, the
new interface in combination with the stable fd guarantee simplifies
the job of the application event loop, and will likely lead to an
overall improvement in application performance.

A XCM socket will now typically use two fds internally, or three fds
in case the control interface is enabled and configured. That is one
more, in the form of the epoll instance fd, than prior to this change.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>