Allow want call on bad TLS connection
A xcm_want() call on a TLS connection in a 'bad' state (e.g. reset or
otherwise defunct) would result in an error being returned. For all
other transports, such a call is successful (with zero fds returned).
In addition, the errno being set would be in violation of the API
(which only allows for EOVERFLOW and EINVAL).

This patch changes TLS transport behavior to be consistent with that
of other transports, and in adherence to the API.