Refactor socket initialization
In order to make the XCM-internal transport socket initialization
easier to understand, the generic part of socket construction happens
in the core library, before the transport's xxx_connect(),
xxx_server(), or xxx_accept() is called (which continues with
transport-specific construction).

The main socket memory structure allocation and deallocation is moved
from the transports to the core library.