Detect changes to XCM_CERT_TLS
Upon establishing a new connection (with xcm_accept() or
xcm_connect)), check the value of the XCM_CERT_TLS environment
variable. This allows single-threaded applications to use the
XCM_CERT_TLS as a crude mechanism to dynamically switch between
different sets of certificate, key and trust chain files.

This patch refactors the SSL context caching mechanism of the TLS
transports, both to accomodate for this new feature, and for future
XCM socket options, which would allow for dynamically selecting
certificate directory.

The number of sets of certificate files is now only limited by
available memory, as opposed to the previous hard limit of 64 sets.