Allow interleaving of reads and writes against OpenSSL
The XCM TLS transport was written under the assumption that in case an
application issued a SSL_read(), and its request could not immediately
be served, it had to repeat the same operation. It was not free to
call SSL_write(). Similarly, SSL_write() operations was repeated in
the face of failure.

In fact, the OpenSSL API is not as limiting, but allows interleaving
non-blocking read and write operations. There is one caveat though:
OpenSSL seemingly has no way to communicate what events to wait for in
case the user wants to wait for both sending and receiving.

This change lifts the "no interleaving" requirement, which simplifies
the TLS transport code. In order to avoid "locking" the connection
into a SSL_read(), there were some syscalls to check if data was
available in the TCP layer. Those are removed, which significantly
reduces run-time overhead in certain situations.

In addition, this fix allows correct operation with OpenSSL 1.1.1f and
later.

Issue ID: 11

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>