Make xcm command-line tool handle buffered messages
The 'xcm' command-line tool didn't properly handle a case when
xcm_want() would return zero fds, signaling that the conditions it's
waiting for (usually XCM_SO_RECEIVABLE) are already met.

With the current XCM transport implementations, combined with the fact
the command-line tool would typically be used for exploration or
debugging, this bug likely hasn't been seen much in the wild.