Allow dynamic update of certificate files
In the XCM TLS transport, before using a cached SSL context, check if
the certificate files has changed. This allows for an external
application to replace the certificate, key and/or trust chain files
being used by running XCM applications. Any new connections will use
the new generation of files. The changes takes effect immediately.

A SHA256 hash of modification time, file size, and dev and inode
number of all files and the directory itself are used to detect
changes. The files need not necessarily be newer than the previous
generation - only different.

Care must be taken to update the directory in an atomic manner, to
avoid a race with XCM application in the process of accepting or
establishing new TLS connections.

Even though this change introduces a number of of fstat() syscalls for
every xcm_accept() or xcm_connect(), it does not noticeably degrade
TLS connection establishment performance.