Updated Python wrapper
The Python XCM API is updated to support the new xcm_await() and
xcm_fd() functions. Since await is a keyword, the Python version of
xcm_await() is called set_target() instead of await(). The xcm_fd()
equivalent is given the more Pythonic name fileno().

The echod.py example program is updated and much simplified. With this
patch, there's no longer a need for a custom event loop, and instead
it uses asyncio.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>