Support writable socket attributes
This change extends the XCM API with a function xcm_attr_set() - an
interface to change the value of socket attribute. Further API
extensions allow supplying a set of attributes at the time of socket
creation. Three new functions have been added for this purpose;
xcm_connect_a(), xcm_server_a(), and xcm_accept_a().

The first use of this framework, also included in this patch, is to
allow applications to choose which local address (and optionally,
port) to be used for originating connections. This is achieved by
supplying an attribute "xcm.local_addr" to xcm_connect_a(). Write
access to this attribute is supported by the TCP, TLS and UTLS
transports. If the port part of the address is set to zero, the kernel
is used for port allocation (just like for in the case of server
sockets).

The XCM_SO_NONBLOCK is backed by a new boolean-type attribute
"xcm.blocking". For reasons of backward compatibility, the flag
XCM_NONBLOCK may still be used in xcm_connect(). xcm_set_blocking()
and xcm_is_blocking() are now just a custom interface to the new
attribute.

The internals of how sockets are constructed have been changed, to
allow setting attributes before a connection has been established or a
server socket bound to its address.

The changes forced a rewrite of the UTLS transport, which in turn
rippled through large amounts code shared among the transports. This
patch moves large chunks of code from xcm.c into xcm_tp.c, to allow it
to be shared by both xcm.c and the UTLS transport. The UTLS
implementation is now done in a much cleaner manner, and does not rely
on the application-level API (i.e. <xcm.h>), but purely uses the
library-internal interface (<xcm_tp.h>).

The UTLS rewrite removes one layer of epoll from the UTLS server
sockets. Prior to this change, the UTLS server socket had its own
epoll instance, and the two subordinate sockets had one each. Now, a
single epoll instance is shared among all three sockets.

The UTLS refactoring does not have any externally-visible effects. In
particular, from the control interface perspective, things look the
same (e.g. the UTLS connection sockets still look like one socket,
although internally they are two).

A side-effect of this change is that the socket id will be incremented
faster, since sockets are created speculatively prior to any
xcm_connect(), xcm_server, or xcm_accept() calls (or their *_a()
variants).

Resolves #4.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>