Fix DNS deadlock on early close
A deadlock may be triggered in a scenario where an application creates
a connection socket with xcm_connect(), the supplied address includes
a DNS name in need of resolution. An immediate xcm_close() call on
such a socket may never return.

This has only been observed when running the XCM test suite in
valgrind, but might well happen in production environments as well.

The bug is triggered in case gai_cancel() is called when a DNS request
has been enqueued to the glibc resolver (via getaddrinfo_a()), but it
has not yet been serviced by glibc's resolver thread. Contrary to the
API specification (as per the manual page), no asynchronous
notification is being sent on EAI_CANCELED. Also, this scenario causes
glibc to leak memory.

There's nothing XCM can do about the memory leak, except accelerate
plans to replace its use of the glibc asynchronous resolver with
something better.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>