Fix unresponsive SCTP socket control interface
For reasons unknown, BSD socket calls on a Linux kernel SCTP socket
may be very time consuming. Since, for performance reasons, the
control interface is not called for every xcm_accept() call, this made
the control interface very slow to respond for SCTP sockets. In some
cases, the time would exceed the libxcmc 300 ms timeout, causing these
sockets' attributes not to show up on 'xcmctl list' (or the
equivalent). It could also cause the failure of the 'xcm:ctl_iter'
test case.

This patch makes the control interface processing function to be
called more frequently on SCTP sockets (only).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>