Add attribute support to the xcm command-line tool
With this patch, the 'xcm' command line program is capable of setting
boolean, integer, and string-type connection attributes.

Binary-type attributes and server sockets attributes are not supported
(since they are not used, with the exception of "xcm.blocking", which
it does not make sense to set from the command line).

This commit also adds a trace entry in case the user attempts to
set a non-existent attribute.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>