Generate test certificates
This patch replaces the hard-coded X.509 certificates and private keys
with a Python cryptography-powered machinery to dynamically generate
per-test case PKI structures. This makes the TLS tests more flexible
and easier to review and extend.

The Python script 'test/gencert.py' generates CA certificates, leaf
certificates and private key files according to a specification (in
YAML format), supplied to it on standard input.

The XCM test suite in turn uses this script to setup an appropriate
structure for each TLS test case.

Resolves #10.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>