Change trusted CA terminology
The file that contained the set of trusted CA certificates was
historically called "trust chain" in the XCM TLS transport
documentation. This term is inaccurate. The trusted certificates are
used to establish a chain of trust, but they do not constitute the
chain of trust.

With this change, the acronym "tc" used for "trust chain" is
retrofitted to be one for "trusted CAs".

This change is targeting humans reading the documentation. For
programs using the API, everything remains the same.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>