X.509 subject name access and verification
This patch introduces support in the TLS transport for verifying the
remote peer's subject names against a name, or set of names. If there
is a non-zero overlap between these two sets, the verification is
considered successful. A "subject name" here is, as per RFC 6125
definition, either a DN of the subject field (in DNS format), or a
subject alternative name extension with the DNS type.

This feature, called "hostname validation" by OpenSSL, is disabled by
default, but may be enabled using the "tls.verify_peer_name" socket
attribute. If enabled, XCM will verify the hostname in the address
supplied in the xcm_connect_a() call. In case the attribute
"tls.peer_names" is also supplied, it overrides this behavior, and
verification will be done against this attribute. The value of the
attribute is a ':'-separated set of subject names. "tls.peer_names"
may be passed to xcm_connect_a(), xcm_server_a(), or xcm_accept_a().

After a TLS has been established, the "tls.peer_names" contains a set
of the remote peer's actual subject names.

The patch bumps the API version 0.17 and the implementation version to
1.3.0.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>