Fix bug in server socket certificate files inheritance
This patch fixes a bug in the TLS transport which caused the
certificate attributes configured on a server socket to override any
certificate attributes set as a part of the xcm_accept_a() call,
leaking memory in the process.

A test case to reproduce the problem is added as well.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>