Improved control processing heuristic
This patch changes the heuristics controlling when XCM will do control
socket processing.

The fds related to the control interface are kept in the same
per-socket epoll instance as the "data" fd (e.g., TCP socket fd).

When the application wakes up due to the XCM socket's fd (i.e., the
epoll instance fd) is active), the API contract mandates a call to a
xcm_*() function on the socket (e.g., xcm_accept()).

Doing control fd processing for every xcm_*() call is too expensive.

Prior to this change, the code would perform control processing every
Nth such call, amortizing the cost across many calls. However, this
scheme has an unfortunate side effect. When the control interface fd
is the reason for the application process waking up (from select() or
similar), it takes on the average N/2 back-to-back wake-ups until
control processing actually occurred (and the control fds are
inactivated). This caused the control interface response to be a
little slow, and, more importantly, caused a load spike when the
control interface was used.

This patches introduces a new approach. First XCM tries the "real"
operation (e.g., an accept() call on the BSD socket fd in case of
xcm_accept()). If and only if this fails with errno set to EAGAIN, XCM
will perform control processing. In this way, control processing
usually only happens when the control fd is active.

To avoid a situation where the control interface is starved (in favor
of actual data fd processing), every 256nd call, the control fds are
processed, regardless to the outcome of the "real" operation.

This change reduces the processing cost and response times related to
the control interface's actual use with ~75%. Other performance
metrics (e.g., xcm_send() and xcm_receive() overhead) are the same or
or slightly improved. An exception is speculatively issued operations,
failing with EAGAIN. An example of such an operation is an application
issuing xcm_receive(), even though it does not know the status of the
socket's epoll fd. Such an operation will often fail, and since
control processing is now always done in the face of a non-fatal
failure (i.e., EAGAIN), the operation is more costly after this patch.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>