Limit the number of epoll instances per active fd
XCM uses shared, always active, event fds to signal to the application
that a socket is active when, for example, a message is buffered or
the connection is closed. Such fds are added to the respective
socket-internal epoll instance, when needed.

The Linux kernel limits the number of epoll instances a file object
can be registered in. See source/fs/eventpoll.c in the kernel for
details.

Before this change, XCM allowed the registration of a single
always-active event fd in an arbitrary number of epoll instances. In
certain situations (e.g, when all connections were lost), where a
process had many (500+) TLS and/or TCP transport connections open, an
epoll_ctl() call which is not supposed to fail with EINVAL did,
triggering an abort.

With this change, XCM grows the number of always-active event fds as
needed, limiting the maximum number of connections using the same
event fd to 100.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>