Disable TLS 1.2 renegotiation
This patch disables TLS 1.2 renegotiation in XCM TLS and BTLS
transports, if built with a recent-enough OpenSSL (1.1.1c or later).

TLS 1.2 renegotiation is considered a potential DoS attack
vector. This protocol feature was never actively used by XCM (i.e. the
XCM library never initiated a renegotiation), and thus can be disabled
without any adverse effects.

Disallowing renegotiation is in line with HTTP/2 use of TLS, as per
section 9.2.1 in RFC 7540.

TLS 1.3 does not support renegotiation, and is not affected by this
change.

Resolves #40.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>