Do not require trusted CA file when TLS authentication is disabled
The TLS and BTLS tranports erroneously required the trusted CA file to
exist and be valid, and then would go on to not use the CA
certificates.

Besides fixing this bug this patch also introduces a TLS/BTLS socket
configuration consistency check that assures that the application
doesn't specify "tls.peer_names" when "tls.verify_peer_name" is set to
false.

Resolves #42.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>