Ignore namespace name lookup failures
The TLS transport family use a set of file system operations
attempting to figure out the name of the current network namespace,
which is used to determine the default certificate, key and trusted CA
file names.

The network namespace code assumed certain /proc files to be in place,
which aren't on non-network namespace capable kernels. This in turn
causes TLS, UTLS and BTLS server and connection socket creation to
fail, unless all necessary certificate related files are given as
socket attributes.

With this patch, network namespace name determination is a best effort
exercise.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>