Disable TLS 1.3 certificate authorities extension
Prior to this patch, when acting as a TLS 1.3 server, XCM would
recommended the remote TLS client which CAs to use by means of the TLS
1.3 certificate_authorities extension. The client, if that also was
the XCM library, will go on to ignore the list provided by the server.

The certificate_authorities extension provides no additional
functionality for XCM, and needlessly discloses information about the
server. This patch discontinue its use by the XCM TLS transport.

This patch does not affect compatibility "on the wire". An application
linked to a XCM library with this change will interoperate with a
remote application using a XCM library without it.

Resolves #48.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>