Remove unnecessary shared object dependencies
Change automake files so that XCM binaries are linked only toward the
shared object libraries they need, as opposed to the union of all
libraries needed by all programs and libraries in the build, which was
the case prior to this patch (except for libevent and libssl).

This change primarily makes the build produce a different result when
such a libtool version is used, that do not remove extranous
dependiences.

The check for libdl is removed, since it is not used.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>