Allow disabling X.509 certificate validity period checks
This patch introduces two new socket attributes
"tls.accept_not_yet_valid" and "tls.accept_expired" which allow an
application to disable checks for the certificate validity period
against the current system time.

The default behavior, which is to reject both not-yet-valid and
expired certificates in the trust chain, and abort TLS connection
establishment, is not changed.

The part of the XCM test suite that handles TLS connection
establishment success/failure in the face of different TLS-related
socket attributes and X.509 certificate properties is refactored to
make room for the new validity period test cases without growing
complexity.

The socket attribute map API is extended with a new utility function
xcm_attr_map_add_all().

Resolves #55.

The API/ABI and implementation versions are bumped to 0.19 and 1.4.1,
respectively.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>