Unify SSL client and server context caches
Merge the two different SSL context caches into one. They only
differed by the verification callback, which could just as well be
explicitly set by the XCM BTLS transport on the SSL object (as opposed
to be inherited from the SSL_CTX).

This patch also paves the way for future changes, where the callback
may be configured differently on a per-socket basis.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>