Merge time check socket attributes
Merge "tls.accept_not_yet_valid" and "tls.accept_expired" XCM TLS
socket attributes into a single attribute "tls.check_time".

This would be a breaking change, but since no release has been made
since this API was introduced, it is not an issue.

The merge greatly simplifies the implementation, at the cost of some
flexibility. With this change, an application cannot express that
expired certificates are OK, but not-yet-valid are not, or vice versa.

The default behavior remains to honor X.509 certificate validity
period.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>