Introduce new version API
This patch introduces a new API xcm_version.h to retrieve API/ABI and
library versions, either at compile time (using the appropriate macros
defines) or at run time (using xcm_version*() function calls).

The 'xcm' tool is updated to use the new API to show XCM library
versioning information on the command-line.

The Python wrapper is extended with two of the xcm_version.h
functions: xcm_version() and xcm_version_api().

This patch bumps the ABI to 0.20 and the library version to 1.5.0.

Resolves #56.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>