Fix rare lack of responiveness on the control interface
Processing of an XCM socket's control interface is performed when the
user application calls xcm_accept(), xcm_send(), xcm_receive(), and
xcm_finish().

This patch fixes a bug which caused control processing to not occur in
the xcm_finish() case, if this function would return EAGAIN.

A socket unable to finish its background tasks for any length of time
is rare, so is likely not triggered often in the wild. During this time,
the XCM socket's fd would be active, likely causing the application
to repeatedly call xcm_finish().

The bug also caused a slight performance degradation in case the errno
variable had a non-zero value at the time of the xcm_finish() call.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>