Add support for IPv6 link local addresses
This patch adds a new integer attribute "ipv6.scope" that may be
passed to xcm_server_a() or xcm_connect_a() to indicate the IPv6 scope
id to used.

The purpose of this attribute to allow the use of IPv6 link local
addresses (i.e., fe80::/10) as a part of a XCM TLS, BTLS or TCP
address.

To reach a server on a IPv6 link local address, a client application
needs, beside the address itself, also know which network interface to
use. This information is conceptually similar to for example which
network namespace to use, which IP interface to bind to (if
source-based routing is employed), or which TLS certificate and keys
to use.

The default scope id is zero (global). On Linux, all other ids are
link local scope ids. Linux employs a direct mapping between scope id
and network interface index (i.e., the index is the scope id). POSIX
if_nametoindex() may be used by applications to translate between
interface name and index, if so is needed.

"ipv6.scope" may not be set on IPv4 sockets.

XCM treats IPv6 scope ids in the same manner as is usally done with
URIs, both in standards and current-day implementations (e.g., libcurl
and recent versions of Firefox).

The IPv6 scope identifiers are strictly local to the node. One node
may use a particular scope id to refer to a particular network, and
another node may use another value for the same network. Thus, the
scope id is arguably not a part of the address.

Another reason for XCM not to adopt the common
[fe80::e011:5dff:fed6:d156%eth0] syntax is that the IP address struct
in <xcm_addr.h> has no room to accommodate the scope id. The struct is
public and thus changing it would break the ABI. This can be worked
around, essentially by introducing a new address parsing API (for the
IP-related functions), and deprecating the old. Needless to say, this
is a big change. Increasing the XCM major version would be an even
more dramatic measure, with even more significant fallout.

This patch also reworks the test suites in relation to test network
configuration (e.g., Linux network namespace management).

ABI version is bumped to 0.21 and implementation version is now 1.6.0.

Resolves #34.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>