Reduce control interface poll frequency
Every time xcm_receive(), xcm_send() or xcm_finish() is called, and
EAGAIN is returned, XCM checks the control interface fd to see if it
(being active) was the reason the application made the XCM call.

This strategy is costly for applications employing speculative
xcm_receive() (e.g., repeatedly calling xcm_receive(), until EAGAIN).
The check is done with a zero-timeout poll() call.

This patch changes this behavior to check the fd only roughly every
four EAGAIN-generating calls. The drawback with this approach is that
if the reason for the XCM call is indeed that the control fd is
active, the application will typically go back into select() (or
equivalent) several times, before the fd is processed. This slows down
the control interface response time and increase the overhead for such
operations.

Earlier versions of XCM implemented a variant of this pattern, but
checked the control fd only every 64 XCM call. Unlike recent XCM, the
counter was pegged for all calls, not only those returning
EAGAIN. This legacy scheme caused excessive CPU usage when the control
interface was used.

Reducing control interface poll frequency as per this change increases
performance with 5-10% for the UX transport in a ping-pong scenario,
and clients and servers using speculative xcm_receive(). Other
transports gain less.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>