Fix bug placing socket in wrong namespace
Prior to this patch, XCM could create a TCP BSD socket in the wrong
network namespace, in case a XCM address with a DNS name was used, and
the application thread changed its network namespace between the
xcm_connect() call, and the follow-up xcm_finish(), xcm_send() or
xcm_receive() call. The API promises that the original network
namespace of the xcm_connect() call will be used for the outgoing
connection.

The TLS, BTLS, TCP, and SCTP transports suffered from this issue.

The root cause was that the libc socket() call was not always made at
the time of xcm_connect(), but later, when DNS resolution had
finished.

With this patch, in case of xcm_connect() calls featuring a DNS
address, two BSD sockets - one AF_INET and one AF_INET6 - will be
created. When DNS resolution has finished, one of the two will be
closed.

This change also includes a test case, reproducing the issue.

Resolves #62.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>