Fix TLS header parsing bug
The TLS transport was unable to properly receive a XCM-level message
with the header split up into two or more TLS protocol-level
records. Such a message triggered an assertion.

The TLS transport always keeps the message header contained in a
single TLS record, but other XCM implementations may choose to do
otherwise.

This bug is a DoS attack vector, but only in case the attacker
possesses the appropriate certificate and private key.

The BTLS transport is not affected by this bug (it does not have a
XCM-level header).

This patch adds a test case which verifies the result of arbitrary,
randomized, mappings of messages to TLS records. It tests both for
cases where multiple messages are carried, in part or in whole, in the
same TLS record, and for scenarios where one XCM message has been
split up across two or more TLS records.

A test case which send random data (e.g., mostly not comforming to the
wire format) on the TLS connection is also added. Such a test case for
the TCP connection level already exists.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>