Add authority key ids to XCM test certificates
The authority key id extension in non-root certificates contains
information on which CA private key was used during for signing it, in
the form of the subject key id.

This extension allows for reliable establishing a chain of trust that
includes one out of multiple CA certificates with the same subject to
be in the same trusted CA bundle. It's mandatory in non-self signed
certificates.

The corresponding test case is updated to reflect the expected
success, when the authority key id is available.

The test root CAs also now has this extension added. In their case
it's their own subject key id that is the value. It serves no
particular function, at least not one that's covered by the XCM test
suite.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>