Add support for configuring TLS credentials by-value
This patch adds three binary attributes on TLS, UTLS and BTLS sockets:
"tls.cert"
"tls.key"
"tls.tc"

The new attributes may be used to pass certificates and related
information by value, rather than by-file-system-reference.

In case an application already has the certificates and keys in
memory, it is more convenient and less error-prone to pass the
information directly to XCM as binary-type XCM socket attributes,
rather than going via a set of temporary files.

The library version is set to 1.7.0 and the ABI is now 0.22.

Resolves #50.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>