Clean up memory exhaustion handling
To somewhat improve code readability, direct abort() calls are
replaced with a call to a new function ut_mem_exhausted(), which in
turn calls abort().

One might question if abort() or exit(EXIT_FAILURE) is the right way
to terminate the process upon encountering an out-of-memory situation.
This patch keeps the legacy behavior (i.e., abort()).

All library instances of abort() calls due to internal errors are
replaced with calls to a new function ut_fatal(), which does abort().

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>