Lower c-ares per-query DNS timeout
Reduce the DNS per-query transaction timeout from c-ares default of 5
seconds, to 1 s. This lower values matches better the typical XCM
application's need for fast response times.

The new timeout value in combination with the c-ares default number of
per-server retries and backoff strategy makes the whole DNS resolution
process for a particular record to time out in ~15 s, provided a
single server is configured, but is not responding.

Resolves #51.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>