Have BTLS transport use BTCP transport
The BTLS transport now delegates all things related to DNS and TCP
byte stream-level communication to the BTCP transport, reducing code
duplication and BTLS complexity.

The new BTLS transport carries a custom OpenSSL BIO, which maps
OpenSSL socket reads and writes to BTCP transport send and receive
calls.

The OpenSSL fd BIO, used by BTCP prior to this commit, did not set the
MSG_NOSIGNAL flag when calling send(). To avoid having the process die
from SIGPIPE when attempting to write on a closed connection, the BTCP
transport used a signal() call to have the SIGPIPE signal ignore.

Ignoring the SIGPIPE is no longer needed (since BTCP does use
MSG_NOSIGNAL), and no longer done.

The BTLS (and TLS) overhead is increased with ~1,5%.

This commit reworks the XCM-internal, per-socket epoll instance
management. The epoll_reg and epoll_reg_set modules are replaced with
a new xpoll module. The xpoll module provides new functionality in the
form of an abstraction for the "always active fd", occasionally used
by various XCM transport implementations to signal that data (or an
error) is available. New functionality is that xpoll allows for
several independent registrations of the same always-active fd, which
is needed to support layered transports (i.e., XCM transports that
delegates part of their work to other transports), several of which
may use the same always-active event fd.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>