Refactor attribute handling
Refactor the XCM-internal transport API so that socket attributes are
not returned as a list, but rather fed back to the caller one-by-one
by means of a callback. Also, remove the odd design with a "context"
parameter to store the actual socket (for cases where one XCM socket
is layered on top of one or more other sockets), and instead
explicitly supply the actual socket pointer.

This change much simplifes the socket attribute handling for composite
transports (i.e., transports which sockets delegates some work to
other transports' sockets).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>