Add support for configurable DNS algorithms
The DNS resolver (glibc or C-ares) sorts the A and AAA records and
return them to the higher layer in the order of preference. XCM
versions prior to this commit only attempted to connect to the first,
most preferred, address in this list. If that connection attempt
failed, an error was returned to the caller.

With this change, the application may control which algorithm is used
to translate the list of addresses into an established TCP connection
using a new XCM socket attribute "dns.algorithm". The default
algorithm is "single", which is identical to legacy XCM behavior.

If the "sequential" method is specified, all addresses provided by DNS
will be probed, one-by-one, in resolver-controlled priority order.

If the "happy_eyeballs" algorithm is configured, the XCM socket
maintains two concurrent tracks, where one attempts to connect to all
IPv6 addresses, one-by-one, in priority order, and the other does the
same, but for the IPv4 subset. All this pretty much in accordance to
RFC 6555. The IPv6 track is given a head start, by delaying the
initial IPv4 connection establishment attempt for 200 ms.

The "dns.algorithm" attribute is available in all TCP-based
transports; TLS, BTLS, TCP, and BTCP. It's also available for
TCP-based UTLS connections. The new socket attribute may be written
only at the time of the xcm_connect_a() call, and may be read only on
XCM connection sockets representing the client side of the TCP
connection.

To keep core BTCP transport complexity low, the responsibility for
managing BSD socket creation and TCP connection establishment is moved
out from xcm_tp_btcp.c to a new module tconnect.c.

This commit only includes test cases for a small part of the new
functionality.

Resolves #37 (except for SCTP).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>