Add support for certificate revocation verification
This patch implements checking the remote peer's certificate (and
other certificates in the chain of trust) against a set of certificate
revocation lists (CRLs).

CRL checking is supported in all XCM transports using TLS, including
for UTLS connection sockets of the TLS type.

CRL checking is disabled by default. It is enabled by setting the new
"tls.check_crl" socket attribute to true.

The actual CRLs are supplied by setting one of the new "tls.crl_file"
or "tls.crl" socket attributes.

The CRL file is cached and re-read on change in the same manner as the
cert.pem, key.pem and tc.pem files. Already-established TLS
connections will not be torn down.

The full chain is checked against the CRLs.

CRL checking is availble on both socket serving in the TLS client and
server roles. It may be configured on both server and connection
sockets, and is inherited from server to connections sockets in the
same manner as for other credential-related pieces of information
(e.g., the private key).

This patch also includes the appropriate updates to the XCM test
suite.

Resolves #77.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>