Stop needlessly requiring trusted CA bundle file
When TLS authentication is disabled, the trusted CA bundle is not
used, and thus should not be required.

Prior to this patch, when the default credential paths were used
(i.e., the key and certificate wasn't specified by means of TLS socket
attributes), the tc.pem file had to be present, even though it was not
used.

Beyond the fix, a test case to reproduce the bug is also included.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>