Fix BTLS OpenSSL BIO bug
Fix bug in BTLS custom OpenSSL BIO which caused SSL_write() to ask the
caller to wait for the BIO to become readable, when in fact it needed
more BIO output buffer space to finish the requested operation, and
thus should ask for the caller to wait for the BIO to become writable.

This bug was introduced when BTLS was moved over from "raw" BSD
sockets to using BTCP (commit fb881f2), and was never present in any
released version of XCM.

This patch reverts the changes introduced by 30923b9, which wasn't
addressing the root cause.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>