Properly set errno on socket address retrieval failures
xcm_local_addr() or xcm_remote_addr() didn't properly set errno in
case of failures (i.e., the address not currently being available).
The same applied for xcm_attr_get() calls to retrieve the same
information.

This patch set the errno to ENOENT in case the address is not
available.

Resolves #84.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>