Properly return empty UX and UXF local and remote address
Starting with XCM v1.9.0, the UX transport's xcm_remote_addr() and
xcm_local_addr() returned NULL in case the peer had failed to set the
SO_PASSCRED socket option.

libxcm sets SO_PASSCRED, but some other implementation may not. XCM
should be able to deal with both situations.

In addition, the UXF transport could return a garbage name, in case
the peer had set SO_PASSCRED on a AF_UNIX using the non-abstract
(i.e., regular pathname-based) naming scheme. XCM sets SO_PASSCRED
on both UX and UXF sockets.

The garbage-name issue was due to a particularity in the Linux kernel,
in where it does not generate (or, autobind as it's called in the
kernel code) a name for a non-abstract AF_UNIX client socket when
SO_PASSCRED is set, but does indicate (via the addrlen parameter) a
6-byte name is present in the sun_path buffer. However, it does not
write to the actual buffer, leaving an unwary app to pick up garbage
to from the stack.

The garbage-name issue has been present for long, but seem to have not
shown up any builds, probably since the stack allocation tended to
include a NUL byte in the sun_path field.

Even though the name was garbage, the result was still NUL-terminated
and thus there was never a risk for a buffer overrun.

This patch fixes both these issues, and adds test cases for both
SO_PASSCRED and non-SO_PASSCRED scenarios.

Resolves #85.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>